@@ -25,18 +25,18 @@
     "android-local": "env-cmd -e dev ionic capacitor run android -address=http://192.168.1.33:8100/"
   },
   "dependencies": {
-    "@awesome-cordova-plugins/core": "^5.39.1",
-    "@awesome-cordova-plugins/media-capture": "^5.39.1",
-    "@awesome-cordova-plugins/native-audio": "^5.39.1",
-    "@capacitor-community/react-hooks": "0.0.11",
-    "@capacitor/android": "3.4.1",
-    "@capacitor/app": "^1.1.0",
-    "@capacitor/camera": "^1.0.1",
-    "@capacitor/core": "^3.4.1",
+    "@awesome-cordova-plugins/core": "^5.44.0",
+    "@awesome-cordova-plugins/media-capture": "^5.44.0",
+    "@awesome-cordova-plugins/native-audio": "^5.44.0",
+    "@capacitor-community/react-hooks": "^0.0.11",
+    "@capacitor/android": "^3.6.0",
+    "@capacitor/app": "^1.1.1",
+    "@capacitor/camera": "^1.3.1",
+    "@capacitor/core": "^4.0.0-beta.1",
     "@capacitor/filesystem": "^1.1.0",
     "@capacitor/haptics": "^1.1.4",
-    "@capacitor/ios": "3.4.1",
-    "@capacitor/keyboard": "^1.2.2",
+    "@capacitor/ios": "^3.6.0",
+    "@capacitor/keyboard": "^1.2.3",
     "@capacitor/status-bar": "^1.0.8",
     "@hookform/error-message": "^2.0.0",
     "@ionic-native/core": "^5.30.0",
@@ -44,9 +44,9 @@
     "@ionic-native/geolocation": "^5.28.0",
     "@ionic-native/media-capture": "^5.36.0",
     "@ionic/pwa-elements": "^3.1.1",
-    "@ionic/react": "^6.0.4",
+    "@ionic/react": "^6.1.12",
     "@ionic/react-hooks": "^0.0.8",
-    "@ionic/react-router": "^6.0.4",
+    "@ionic/react-router": "^6.1.12",
     "@react-google-maps/api": "^1.9.7",
     "@reduxjs/toolkit": "^1.4.0",
     "@stripe/react-stripe-js": "^1.4.1",
@@ -69,10 +69,10 @@
     "jodit-react": "^1.0.66",
     "php-serialize": "^4.0.2",
     "precise-ui": "^1.5.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-audio-player": "^0.13.0",
     "react-cropper": "^2.1.7",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-hook-form": "^7.18.1",
     "react-image-crop": "^8.6.4",
     "react-media-recorder": "^1.6.4",
@@ -92,7 +92,7 @@
     "videojs-record": "^4.5.0"
   },
   "devDependencies": {
-    "@capacitor/cli": "^3.4.1",
+    "@capacitor/cli": "^4.0.0-beta.1",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.3",
     "@testing-library/user-event": "^13.5.0",
@@ -105,6 +105,7 @@
     "@types/react-router-dom": "^5.1.3",
     "@types/video.js": "^7.3.10",
     "eslint-plugin-react-hooks": "^4.2.0",
+    "postcss-import": "^14.1.0",
     "react-error-overlay": "6.0.9",
     "react-scripts": "^4.0.3",
     "redux-logger": "^3.0.6",
