@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from "react-dom/client";
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 
@@ -34,7 +34,6 @@ const store = createStore();
 
 axios.defaults.baseURL = process.env.REACT_APP_API_BASE_URL; // Beta
 // axios.defaults.baseURL = 'http://localhost:8888/LocalFirst/trunk/rest/';
-axios.defaults.baseURL = 'https://www.onagon.com/rest/';
 //axios.defaults.headers.common['Authorization'] = 'AUTH TOKEN';
 axios.defaults.headers.post['Content-Type'] = 'application/json';
 // axios.defaults.headers.post['Access-Control-Allow-Origin'] = '*';
@@ -63,10 +62,16 @@ axios.interceptors.response.use((response: any) => {
     return Promise.reject(error);
 });
 
-ReactDOM.render(
-<Provider store={store}>
-    <App />
-</Provider>, document.getElementById('root'));
+const container = document.getElementById('root');
+const root = createRoot(container);
+
+root.render(
+    <React.StrictMode>
+        <Provider store={store}>
+            <App />
+        </Provider>
+    </React.StrictMode>
+);
 
 // Call the element loader after the platform has been bootstrapped
 defineCustomElements(window);  // add this line
