@@ -55,10 +55,10 @@ const MediaList: React.FC<PropsInterface> = (props: PropsInterface) => {
                 }
                 if( item.uploaded_name && (+authUser.ID === +item.mem_id || +item.converted === 1)) {
                     str = <IonItem className="cursor" lines="none" key={index} onClick={() => categoryModalFn(`${title} Viewer`, item.mem_id, item.rep_id, item.form_id, item.upload_type, item.filename, item.converted)} >
-                        <IonAvatar slot="start" color="greenbg">
-                            { item && ['document', 'article'].includes(item.upload_type) && <IonIcon className="pt-2" color="greenbg" size="large" icon={documentTextOutline}></IonIcon>}
-                            { item && item.upload_type === 'audio' && <IonIcon className="pt-2" color="greenbg" size="large" icon={musicalNoteOutline}></IonIcon>}
-                            { item && item.upload_type === 'video' && <IonIcon className="pt-2" color="greenbg" size="large" icon={videocamOutline}></IonIcon>}
+                        <IonAvatar slot="start" color="appbg">
+                            { item && ['document', 'article'].includes(item.upload_type) && <IonIcon className="pt-2" color="appbg" size="large" icon={documentTextOutline}></IonIcon>}
+                            { item && item.upload_type === 'audio' && <IonIcon className="pt-2" color="appbg" size="large" icon={musicalNoteOutline}></IonIcon>}
+                            { item && item.upload_type === 'video' && <IonIcon className="pt-2" color="appbg" size="large" icon={videocamOutline}></IonIcon>}
                         </IonAvatar>
                         <IonLabel>
                             <h2>{item.upload_title? item.upload_title: item.uploaded_name}</h2>
