@@ -80,7 +80,7 @@ const DeleteModal: React.FC<Props> = ({ title, showDeleteModal, setShowDeleteMod
   return (<>
     <form onSubmit={handleSubmit(onSubmitFn)}>
       <IonHeader translucent>
-        <IonToolbar color="greenbg">
+        <IonToolbar color="appbg">
             <IonButtons slot={ isPlatform('desktop')? 'end': 'start' }>
                 <IonButton onClick={() => setShowDeleteModal({isOpen: false, id: null, mem_id: null})}>
                     <IonIcon icon={close} slot="icon-only"></IonIcon>
@@ -132,7 +132,7 @@ const DeleteModal: React.FC<Props> = ({ title, showDeleteModal, setShowDeleteMod
               
               <div className="mt-4">           
               { (isPlatform('desktop')) && 
-                <IonButton color="greenbg" className="ion-margin-top mt-4 mb-3 float-right" type="submit" >
+                <IonButton color="appbg" className="ion-margin-top mt-4 mb-3 float-right" type="submit" >
                     Submit
                 </IonButton>
               }
